requirements.txt
opencv-python
cupy-12x
scipy
joblib